requirements.txt
tk
pyautogui
pillow
google-genai
langchain
langchain-groq
groq
langchain-core
PyPDF2
python-dotenv
pymongo
reportlab
